+++ library-collection/src/main/java/universum/studios/android/preference/ArrayPreference.java
@@ -67,7 +67,7 @@ public final class ArrayPreference<T> extends SharedPreference<T> {
 	}
 
 	/**
-	 * <b>This constructor has been deprecated and will be removed in the next release.</b>
+	 * <b>This constructor has been deprecated and will be removed in the next none-beta release.</b>
 	 * <p>
 	 * Creates a new instance of ArrayPreference.
 	 *