+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Change-Log
 ===============
 
+### [Release 2.0.0-beta2](https://github.com/universum-studios/android_preferences/releases/tag/2.0.0-beta2) ###
+> 05.04.2017
+
+- Refactored implementation of `PreferencesManager`. Now only implements `SharedPreferencesFacade`
+  along with `SharedPreferencesHost` and may be created via simple constructor taking only `Context`
+  as parameter. Desired configuration related to preferences file name and mode may be specified via
+  `setSharedPreferencesName(String)` and `setSharedPreferencesMode(int)`.
+- Refactored implementation of `SimpleSharedPreferencesFacade`. Instance of simple facade may be now
+  created via simple constructor taking only instance of `SharedPreferences` as parameter.
+
 ### [Release 2.0.0-beta1](https://github.com/universum-studios/android_preferences/releases/tag/2.0.0-beta1) ###
 > 02.04.2017